Move animation package
