+++ trikot-viewmodels-declarative/viewmodels-declarative/src/commonMain/kotlin/com/mirego/trikot/viewmodels/declarative/animation/VMDAnimation.kt
@@ -1,4 +1,4 @@
-package com.mirego.trikot.viewmodels.declarative.components.animation
+package com.mirego.trikot.viewmodels.declarative.animation
 
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
