+++ trikot-viewmodels-declarative/viewmodels-declarative/src/commonMain/kotlin/com/mirego/trikot/viewmodels/declarative/animation/VMDAnimationContext.kt
@@ -1,4 +1,4 @@
-package com.mirego.trikot.viewmodels.declarative.components.animation
+package com.mirego.trikot.viewmodels.declarative.animation
 
 import com.mirego.trikot.foundation.concurrent.AtomicStackReference
