+++ trikot-viewmodels-declarative/viewmodels-declarative/src/commonMain/kotlin/com/mirego/trikot/viewmodels/declarative/animation/VMDAnimationEasing.kt
@@ -1,4 +1,4 @@
-package com.mirego.trikot.viewmodels.declarative.components.animation
+package com.mirego.trikot.viewmodels.declarative.animation
 
 sealed interface VMDAnimationEasing {
     enum class Standard: VMDAnimationEasing {
