+++ trikot-viewmodels-declarative/viewmodels-declarative/src/commonMain/kotlin/com/mirego/trikot/viewmodels/declarative/viewmodel/internal/VMDPropertyChange.kt
@@ -1,6 +1,6 @@
 package com.mirego.trikot.viewmodels.declarative.viewmodel.internal
 
-import com.mirego.trikot.viewmodels.declarative.components.animation.VMDAnimation
+import com.mirego.trikot.viewmodels.declarative.animation.VMDAnimation
 import kotlin.reflect.KProperty
 
 class VMDPropertyChange<V>(val property: KProperty<V>, val oldValue: V, val newValue: V, val animation: VMDAnimation? = null)
